refactor dataclasses into `_workflow_instance.py`; use `typing` package to ensure compatibility.
