+++ temporalio/common.py
@@ -1044,35 +1044,3 @@ def _type_hints_from_func(
         # necessarily
         args.append(arg_hint)  # type: ignore
     return args, ret
-
-
-@dataclass
-class _SDKInfo:
-    name: str
-    version: str
-
-
-@dataclass
-class _FileSlice:
-    content: str
-    lineOffset: int
-
-
-@dataclass
-class _FileLocation:
-    filePath: str
-    line: Optional[int] = -1
-    column: Optional[int] = -1
-    functionName: Optional[str] = None
-
-
-@dataclass
-class _StackTrace:
-    locations: list[_FileLocation]
-
-
-@dataclass
-class _EnhancedStackTrace:
-    sdk: _SDKInfo
-    sources: dict[str, _FileSlice]
-    stacks: list[_StackTrace]
